Change facebook function to service user
